Simplify guards for array instantiation
Previously, we checked various flags, which is more flexible since it allows complex subclassing.

Now the guards are simplified to check for the known class. This should be ok for now, but will break if we subclass further.

In that case, we should try a caching approach to keep a simple pointer comparison for the check.

Signed-off-by: Stefan Marr <git@stefan-marr.de>